% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/house_animals.R
\docType{data}
\name{house_animals}
\alias{house_animals}
\title{Data Frame of House Animals with their Sounds}
\format{
A data frame with 2 rows and 2 variables:
\describe{
\item{animal}{The animal as a character string.}
\item{sound}{The corresponding sound.}
}
}
\usage{
house_animals
}
\description{
A dataset of animals (currently cats and dogs) from the home
and their sounds.
}
\keyword{datasets}
