% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Animal Sounds Function - all your favourite animals!}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{sound}{}
}
\description{
This function takes two character strings. One for
the animal you want to represent and one for the
sound it makes. There's an easter egg for entering
"fox". We use the \code{paste0} function in our script.
To find out more, visit the \link[base:paste]{documentation}.
We have not used \code{\link[basemodels:dummy_classifier]{basemodels::dummy_classifier()}}.
}
